% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_count.R
\name{spark_str_count}
\alias{spark_str_count}
\title{Count the number of matches in a Spark DataFrame's column.}
\usage{
spark_str_count(base, column, pattern)
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}

\item{pattern}{A pattern that must be counted.}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
Count the number of matches in a Spark DataFrame's column.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_count(iris_tbl, 'Species', 'a')
}
