% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_detect.R
\name{spark_str_detect}
\alias{spark_str_detect}
\title{Detect the presence or absence of a pattern in a Spark DataFrame's column.}
\usage{
spark_str_detect(base, column, pattern)
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}

\item{pattern}{Pattern to look for.}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
Detect the presence of a pattern match in a Spark DataFrame's column.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_detect(iris_tbl, 'Species', 'a')
}
