% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_extract.R
\name{spark_str_extract}
\alias{spark_str_extract}
\title{Return the first pattern match found in each string, as a Spark DataFrame's column.}
\usage{
spark_str_extract(base, column, pattern)
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}

\item{pattern}{A pattern to look for.}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
Return the first pattern match found in each string, as a Spark DataFrame's column.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_extract(iris_tbl, 'Species', 'a')
}
