% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_length.R
\name{spark_str_length}
\alias{spark_str_length}
\title{The width of string in a Spark DataFrame's column.}
\usage{
spark_str_length(base, column)
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
The width of string (i.e. number of code points, which generally equals the number of characters) in a Spark DataFrame's column.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_length(iris_tbl, 'Species')
}
