% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_match.R
\name{spark_str_match}
\alias{spark_str_match}
\title{Extract matched groups from a Spark DataFrame's column.}
\usage{
spark_str_match(base, column, pattern)
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}

\item{pattern}{Pattern to look for.}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
Return the first pattern match found in each string, originally as a matrix with a column for each () group in pattern. \code{spark_str_extract} returns the result as a Spark DataFrame's column.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_match(iris_tbl, 'Species', 'a')
}
