% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_pad.R
\name{spark_str_pad}
\alias{spark_str_pad}
\title{Pad a string from a Spark DataFrame's column.}
\usage{
spark_str_pad(
  base,
  column,
  width,
  side = c("left", "right", "both"),
  pad = " "
)
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}

\item{width}{Minimum width of padded strings.}

\item{side}{Side on which padding character is added (left, right or both).}

\item{pad}{Single padding character (default is a space).}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
Pad string in a Spark DataFrame's column to constant width.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_pad(iris_tbl, 'Species', 20, 'right', 'a')
}
