% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_trim.R
\name{spark_str_trim}
\alias{spark_str_trim}
\title{Trim whitespace in string from a Spark DataFrame's column.}
\usage{
spark_str_trim(base, column, side = c("both", "left", "right"))
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}

\item{side}{Side on which to remove whitespace (left, right or both).}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
Trim whitespace from start and end of string in a Spark DataFrame's column.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_trim(iris_tbl, 'Species', 'right')
}
