% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_str_trunc.R
\name{spark_str_trunc}
\alias{spark_str_trunc}
\title{Truncate a character string from a Spark DataFrame's column.}
\usage{
spark_str_trunc(
  base,
  column,
  width,
  side = c("right", "left", "center"),
  ellipsis = "..."
)
}
\arguments{
\item{base}{A Spark DataFrame.}

\item{column}{A column in the Spark DataFrame.}

\item{width}{Maximum width of string.}

\item{side, ellipsis}{Location and content of ellipsis that indicates content has been removed.}
}
\value{
A new column called 'result' in the same Spark DataFrame with the changes desired upon the column given previously.
}
\description{
Truncate a character string in a Spark DataFrame's column.
}
\examples{
iris_tbl <- copy_to(sc, iris, overwrite = TRUE)

spark_str_trunc(iris_tbl, 'Species', 5, 'right', '...')
}
